fix(x64,client): proper warning fixes + remove CRT/truncation suppressions + 64-bit handle pipeline
Compiles clean (0 compile errors; only the cryptlib link path remains for the
link phase). Drives the x64 client toward a genuinely warning-clean, truncation-
safe state with proper fixes, never suppression.

- Per-file proper fixes across 97 files (size_t->size_t/explicit cast, pointer<->int
  via uintptr_t, format specifiers ->%lld for widened GoldType/TimeT64, CRT->_s
  bounded variants). Central: C4652 (ScriptLib PCH browse-info), C5055 (enum/float
  cast), C4566 (Util.cpp CJK literal -> explicit GBK bytes), SnowEnvironment min
  macro, DXTCImage _strupr_s.
- Remove ALL _CRT_SECURE_NO_WARNINGS (4 StdAfx.h + 16 vcxproj) and the
  #pragma warning(disable:4244/4018/4245) that globally hid x64 truncation and
  signed/unsigned warnings across the client. Kept only cosmetic L4 pragmas
  (4710/4786/4512/4201) that don't fire at /W3 and/or originate in DirectX/Windows
  SDK headers. Surfaced warnings are being fixed, not re-hidden.
- 64-bit-safe Python handle pipeline: pointers were round-tripped through Python as
  32-bit ints (Py_BuildValue("i",ptr) + PyTuple_GetInteger->int), truncating any
  pointer above 4GB on x64. New central helpers Py_BuildHandle / PyTuple_GetHandle
  in ScriptLib/PythonUtils carry the full pointer width; migrated 37 writes + 17
  reads across the Graphic/Window/Item/Skill handle modules. Genuine-int returns
  left intact.
- Light-version key: widened the ApplyLight / UpdateTerrainLighting / m_dwVersion /
  m_dwLightVersion chain DWORD->DWORD_PTR (MapBase/MapOutdoor/MapOutdoorRender/
  TerrainPatch/MapManager) so the pointer-derived version key keeps full identity.

+++ Lead-Client-Source/EffectLib/EffectMesh.cpp
@@ -32,7 +32,7 @@ DWORD CEffectMesh::GetFrameCount()
 
 DWORD CEffectMesh::GetMeshCount()
 {
-	return m_pEffectMeshDataVector.size();
+	return static_cast<DWORD>(m_pEffectMeshDataVector.size());
 }
 
 CEffectMesh::TEffectMeshData * CEffectMesh::GetMeshDataPointer(DWORD dwMeshIndex)
@@ -187,7 +187,7 @@ BOOL CEffectMesh::__LoadData_Ver002(int iSize, const BYTE * c_pbBuf)
 		pMeshData->pImageVector.clear();
 
 		std::string strExtension;
-		GetFileExtension(pMeshData->szDiffuseMapFileName, strlen(pMeshData->szDiffuseMapFileName), &strExtension);
+		GetFileExtension(pMeshData->szDiffuseMapFileName, static_cast<int>(strlen(pMeshData->szDiffuseMapFileName)), &strExtension);
 		stl_lowers(strExtension);
 
 		if (0 == strExtension.compare("ifl"))
@@ -332,7 +332,7 @@ BOOL CEffectMesh::__LoadData_Ver001(int iSize, const BYTE * c_pbBuf)
 		pMeshData->pImageVector.clear();
 
 		std::string strExtension;
-		GetFileExtension(pMeshData->szDiffuseMapFileName, strlen(pMeshData->szDiffuseMapFileName), &strExtension);
+		GetFileExtension(pMeshData->szDiffuseMapFileName, static_cast<int>(strlen(pMeshData->szDiffuseMapFileName)), &strExtension);
 		stl_lowers(strExtension);
 
 		if (0 == strExtension.compare("ifl"))
@@ -497,7 +497,7 @@ bool CEffectMeshScript::GetMeshDataPointer(DWORD dwMeshIndex, TMeshData ** ppMes
 
 int CEffectMeshScript::GetMeshDataCount()
 {
-	return m_MeshDataVector.size();
+	return static_cast<int>(m_MeshDataVector.size());
 }
 
 int CEffectMeshScript::GetBillboardType(DWORD dwMeshIndex)
