+++ Lead-Client-Source/EterBase/Poly/Poly.cpp
@@ -579,7 +579,7 @@ int CPoly::SetVar(const std::string & strName, double dVar)
 
 int CPoly::GetVarCount()
 {
-	return lSymbol.size() - MathSymbolCount;
+	return static_cast<int>(lSymbol.size() - MathSymbolCount);
 }
 
 const char * CPoly::GetVarName(unsigned int dwIndex)
