+++ Lead-Client-Source/EterBase/Utils.cpp
@@ -180,7 +180,7 @@ void GetOnlyFileName(const char * sz_Name, std::string & strFileName)
 	strFileName = "";
 
 	int i;
-	for (i=strlen(sz_Name)-1; i>=0; --i)
+	for (i=static_cast<int>(strlen(sz_Name))-1; i>=0; --i)
 	{
 		if ('\\' == sz_Name[i] || '/' == sz_Name[i])
 		{
@@ -188,7 +188,7 @@ void GetOnlyFileName(const char * sz_Name, std::string & strFileName)
 			break;
 		}
 	}
-	
+
 	if (i == -1)
 		i = 0;
 
@@ -208,7 +208,7 @@ void GetExceptionPathName(const char * sz_Name, std::string & OnlyFileName)
 void GetOnlyPathName(const char * sz_Name, std::string & OnlyPathName)
 {
 	int i;
-	for (i = strlen(sz_Name) - 1; i >= 0; --i)
+	for (i = static_cast<int>(strlen(sz_Name)) - 1; i >= 0; --i)
 	{
 		if ('\\' == sz_Name[i] || '/' == sz_Name[i])
 		{
@@ -250,7 +250,7 @@ bool GetLocalFileName(const char * c_szGlobalPath, const char * c_szFullPathFile
 	if (strGlobalPath.length() >= strFullPathFileName.length())
 		return false;
 
-	DWORD length = min(strGlobalPath.length(), strFullPathFileName.length());
+	DWORD length = static_cast<DWORD>(min(strGlobalPath.length(), strFullPathFileName.length()));
 	for (DWORD dwPos = 0; dwPos < length; ++dwPos)
 	{
 		if (strGlobalPath[dwPos] != strFullPathFileName[dwPos])
@@ -439,7 +439,7 @@ void MyCreateDirectory(const char* path)
 
 	p = path;
 
-	int len = strlen(path) + 1;
+	size_t len = strlen(path) + 1;
 	dir = new char[len];
 
 	while (*p)
@@ -526,7 +526,7 @@ void StringExceptCharacter(std::string * pstrString, const char * c_szCharacter)
 	int icurPos = 0;
 	int iNextPos = 0;
 
-	while((iNextPos = pstrString->find_first_of(c_szCharacter, icurPos)) >= 0)
+	while((iNextPos = static_cast<int>(pstrString->find_first_of(c_szCharacter, icurPos))) >= 0)
 	{
 		std::string strFront = pstrString->substr(icurPos, iNextPos - icurPos);
 		std::string strBack = pstrString->substr(iNextPos+1, pstrString->length() - iNextPos - 1);
@@ -545,7 +545,7 @@ bool SplitLine(const char * c_szLine, const char * c_szDelimeter, std::vector<st
 
 	do
 	{
-		int beginPos = strLine.find_first_not_of(c_szDelimeter, basePos);
+		int beginPos = static_cast<int>(strLine.find_first_not_of(c_szDelimeter, basePos));
 		if (beginPos < 0)
 			return false;
 
@@ -554,7 +554,7 @@ bool SplitLine(const char * c_szLine, const char * c_szDelimeter, std::vector<st
 		if (strLine[beginPos] == '"')
 		{
 			++beginPos;
-			endPos = strLine.find_first_of("\"", beginPos);
+			endPos = static_cast<int>(strLine.find_first_of("\"", beginPos));
 
 			if (endPos < 0)
 				return false;
@@ -563,7 +563,7 @@ bool SplitLine(const char * c_szLine, const char * c_szDelimeter, std::vector<st
 		}
 		else
 		{
-			endPos = strLine.find_first_of(c_szDelimeter, beginPos);
+			endPos = static_cast<int>(strLine.find_first_of(c_szDelimeter, beginPos));
 			basePos = endPos;
 		}
 
@@ -608,7 +608,7 @@ PCHAR* CommandLineToArgv( PCHAR CmdLine, int* _argc )
 	BOOLEAN  in_TEXT;
 	BOOLEAN  in_SPACE;
 
-	len = strlen(CmdLine);
+	len = static_cast<ULONG>(strlen(CmdLine));
 	i = ((len+2)/2)*sizeof(PVOID) + sizeof(PVOID);
 
 	argv = (PCHAR*)GlobalAlloc(GMEM_FIXED,
